fialing on ensure namespace
